+++ lib/axios.js
@@ -11,7 +11,6 @@ const token = Cookies.get('token')
 const headers = {
   'Accept': 'application/json',
   'Content-Type': 'application/json',
-  // 'Connection' : 'keep-alive'
 }
 
 if (token) {
